Result Pattern

+++ Application/Service/DocumentService.cs
@@ -1,10 +1,13 @@
-﻿using Application.Dto;
+﻿using Application.Common;
+using Application.Dto;
 using Application.Interface;
 using AutoMapper;
 using Domain.Interface;
 using Domain.Models;
 using Microsoft.AspNetCore.Hosting;
+using System.Collections.Generic;
 using System.Reflection.Metadata;
+using System.Xml.Linq;
 
 namespace Application.Service
 {
@@ -19,20 +22,25 @@ namespace Application.Service
             _mapper = mapper;
         }
 
-        public async Task<bool> CreateDocument(DocumentUploadDto document, int userId)
+        public async Task<Result> CreateDocument(DocumentUploadDto document, int userId)
         {
+            if (document.File == null || document.File.Length == 0)
+            {
+                return Result.Failure("File not provided");
+            }
+
             if (document == null)
-                return false;
+                return Result.Failure("document not found");
 
             var folder = await _unitOfWork.FolderRepository.GetFolderById(document.FolderId, userId);
 
             if (folder == null)
-                return false;
+                return Result.Failure("folder not found");
 
             var workspace = await _unitOfWork.WorkspaceRepository.GetWorkspaceByUser(userId);
             if (workspace == null)
             {
-                return false;
+                return Result.Failure("workspace not found");
             }
 
 
@@ -61,68 +69,105 @@ namespace Application.Service
 
             var newDocument = _mapper.Map<Domain.Models.Document>(documentDto);
             await _unitOfWork.DocumentRepository.Add(newDocument);
-            return _unitOfWork.Save() > 0;
-
-
+            var result = _unitOfWork.Save() > 0;
+            if (result)
+                return Result.Success();
+            else
+                return Result.Failure("Falied to create document");
         }
 
-        public async Task<bool> DeleteDocument(int id, int userId)
+        public async Task<Result> DeleteDocument(int id, int userId)
         {
+            if (id < 0)
+                return Result.Failure("Invalid Id");
+
             var document = await _unitOfWork.DocumentRepository.GetDocumentByUser(userId, id);
             if (document != null)
                 _unitOfWork.DocumentRepository.Delete(document);
 
-            return _unitOfWork.Save() > 0 ? true : false;
+            var result = _unitOfWork.Save() > 0;
+            if (result)
+                return Result.Success();
+            else
+                return Result.Failure("Falied to delete document");
         }
 
-        public async Task<DocumentDto> GetDocumentMetadata(int id, int userId)
+        public async Task<GenericResult<DocumentDto>> GetDocumentMetadata(int id, int userId)
         {
+            if (id < 0)
+                return GenericResult<DocumentDto>.Failure("Invalid Id");
+
             var document = await _unitOfWork.DocumentRepository.GetDocumentByUser(userId, id);
-            return _mapper.Map<DocumentDto>(document);
+            if (document == null)
+            {
+                return GenericResult<DocumentDto>.Failure("Document not found");
+            }
+            var documentDto = _mapper.Map<DocumentDto>(document);
+            return GenericResult<DocumentDto>.Success(documentDto);
         }
 
-        public async Task<IEnumerable<DocumentDto>> GetDocumentOfWorkspace
+        public async Task<GenericResult<IEnumerable<DocumentDto>>> GetDocumentsOfWorkspace
             (int userId, int workspaceId, string documentName = null, string documentType = null, DateTime? creationDate = null)
         {
+            if (workspaceId < 0)
+                return GenericResult<IEnumerable<DocumentDto>>.Failure("Invalid Workspace Id");
+
             var documents = await _unitOfWork.DocumentRepository.GetDocumentsByWorkspace(userId, workspaceId, documentName, documentType, creationDate);
-            return _mapper.Map<List<DocumentDto>>(documents);
-
+            if(documents == null)
+            {
+                return GenericResult<IEnumerable<DocumentDto>>.Failure("Documents not found");
+            }
+            var documentDto = _mapper.Map<List<DocumentDto>>(documents);
+            return GenericResult<IEnumerable<DocumentDto>>.Success(documentDto);
         }
 
-        public async Task<string> GetDocumentPath(int id, int userId)
+        public async Task<GenericResult<string>> GetDocumentPath(int id, int userId)
         {
-            return await _unitOfWork.DocumentRepository.GetDocumentPath(id, userId);
+            if (id < 0)
+                return GenericResult<string>.Failure("Invalid Id");
+
+            var path = await _unitOfWork.DocumentRepository.GetDocumentPath(id, userId);
+            if (path == null)
+                return GenericResult<string>.Failure("Can't find path");
+            else
+                return GenericResult<string>.Success(path);
         }
 
-        public async Task<IEnumerable<DocumentDto>> GetDocumentsByFolder(int FolderId, int userId)
+        public async Task<GenericResult<IEnumerable<DocumentDto>>> GetDocumentsByFolder(int FolderId, int userId)
         {
+            if (FolderId < 0)
+                return GenericResult<IEnumerable<DocumentDto>>.Failure("Invalid Folder Id");
+
             var documents = await _unitOfWork.DocumentRepository.GetDocumentsByFolder(FolderId, userId);
-            return _mapper.Map<List<DocumentDto>>(documents);
+            if (documents == null)
+            {
+                return GenericResult<IEnumerable<DocumentDto>>.Failure("Documents not found");
+            }
+            var documentDto = _mapper.Map<List<DocumentDto>>(documents);
+            return GenericResult<IEnumerable<DocumentDto>>.Success(documentDto);
         }
 
-        public async Task<(bool isDownloaded , string message , byte[] fileBytes , string mimiType , string documentName)> DownloadDocument(int id, int userId)
+        public async Task<GenericResult<(byte[] fileBytes, string mimiType, string documentName)>> DownloadDocument(int id, int userId)
         {
             var document = await _unitOfWork.DocumentRepository.GetDocumentToDownload(id, userId);
 
             if (document == null)
             {
-                return (false ,"Document Not Found", Array.Empty<byte>(), "","");
+                return GenericResult<(byte[] fileBytes, string mimiType, string documentName)>.Failure("Document Not Found")  ;
             }
-            string filePath = await GetDocumentPath(id, userId);
+            var path = await GetDocumentPath(id, userId);
+            string filePath = path.Value;
 
             if (!System.IO.File.Exists(filePath))
             {
-                return (false,"File Not Found", Array.Empty<byte>(), "", "");
+                return GenericResult<(byte[] fileBytes, string mimiType, string documentName)>.Failure("File Not Found");
             }
             var fileBytes = await System.IO.File.ReadAllBytesAsync(filePath);
             var mimiType = GetMimeType(document.Type);
 
-            return (true, "Document downloaded sucessfully",fileBytes,mimiType,document.Name);
-
-
-
-
+            return GenericResult<(byte[] fileBytes, string mimiType, string documentName)>.Success((fileBytes, mimiType, document.Name));
         }
+
         public string GetMimeType(string fileExtension)
         {
             return fileExtension switch
@@ -135,23 +180,34 @@ namespace Application.Service
             };
         }
 
-        public async Task<bool> UpdateDocumentName(int documentId, string name, int userId)
+        public async Task<Result> UpdateDocumentName(int documentId, string name, int userId)
         {
+            if (documentId < 0)
+                return Result.Failure("Invalid Id");
+
+            if (string.IsNullOrEmpty(name))
+                return Result.Failure("Name not Entered");
+
             var document = await _unitOfWork.DocumentRepository.GetDocumentByUser(userId, documentId);
             if (document == null)
-                return false;
+                return Result.Failure("document not found");
             
             var folderPath = GetFolderPath(document.Folder.Workspace.Name, document.Folder.Name);
             var oldPath = Path.Combine(folderPath, document.Name);
             var newPath = Path.Combine(folderPath, name);
 
             document.Name = $"{name}.{document.Type}";
+            document.Path = $"{newPath}.{document.Type}";
 
             if (!renameFile(oldPath, newPath,document.Type))
-                return false;
+                return Result.Failure("Falied to update document");
 
             _unitOfWork.DocumentRepository.Update(document);
-            return _unitOfWork.Save() > 0;
+            var result = _unitOfWork.Save() > 0;
+            if (result)
+                return Result.Success();
+            else
+                return Result.Failure("Falied to update document"); 
 
         }
